\name{summary.irates}
\alias{summary.irates}
\title{
Summary method for 'irates' object
}
\description{
Summary method for an object of class \code{irates}. It prints all competing events incidence rates contained in \code{irates}, including full sample incidence rates. It also prints the baseline covariate characteristics.
}
\usage{
\S3method{summary}{irates}(object, ...)
}
\arguments{
  \item{object}{
\code{irates} object
}
  \item{\dots}{
Further arguments}
}
\value{
No value is returned
}
\seealso{
\code{\link{irates}}, \code{\link{print.irates}}
}
